fix: `Dockerfile` - Remove `build` stage (`setcap`)
When the image was migrated to run as a non-root user some there was a common issue requiring the capability `NET_BIND_SERVICE` when unprivileged ports are set to start at 1024 (host default), such as for port 53 within the container.

To remedy this concern `setcap cap_net_bind_service=+ep /coredns` was used. Upon release of the new CoreDNS image it provided some convenience but it also caused breakage for another set of users, notably those following best practices to lock down capabilities that are not required by their deployments at runtime (_such as binding to an unprivileged port instead, where in some environments the privilege ports range was implicitly or explicitly lowered to `0` via the associated sysctl `net.ipv4.ip_unprivileged_port_start`_).

While UX convenience is preferable it should not be at the expense of denying valid security practices. In addition to this difference is the error messaging communicated upon failure, without `setcap` enforcing the capability as mandatory it is communicated clearly that the failure was due to a lack of permission to bind the desired port; whereas with the `setcap` workaround the kernel enforced check for the capability occurs before the program can execute, with a vague failure of `operation not permitted`.

Today in 2025 with the adoption of containerd 2.0 with it's CRI plugin defaulting the sysctl `net.ipv4.ip_unprivileged_port_start` to `0` like Podman and Docker, much of the ecosystem does not require relying on `setcap` here for the convenience it provided for non-root process to require the `NET_BIND_SERVICE`  capability.

Those that are still running outdated environments can still explicitly opt-in to lowering the associated sysctl for their environment, thus it is no more a breaking change than when `setcap` itself was introduced. The `setcap` solution itself has been known to fail in several deployment environments where that information was lost (_effectively becoming the unmodified release of CoreDNS distributed elsewhere outside of the official image_).

Signed-off-by: Brennan Kinney <5098581+polarathene@users.noreply.github.com>